+++ sdk/spring/azure-spring-integration-servicebus/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release History
 
 ## 2.0.0-beta.2 (Unreleased)
+### Breaking Changes
+- Deprecated the `spring.cloud.azure.managed-identity.client-id` property,
+  use `spring.cloud.azure.client-id` to set the managed identity id when using Managed Identity.
 
 ## 2.0.0-beta.1 (2020-11-18)
 ### Breaking Changes